+++ examples/credential-registry-storage-contract/src/lib.rs
@@ -1,25 +1,23 @@
 //! The contract is used for storing verifiable credentials for the Web3Id
 //! infrastructure. The contract stores encrypted credentials (generated with
-//! the symmetric AES-256 encryption standard) with some metadata associated
-//! with them.
+//! the symmetric AES-256 encryption standard) and a version.
 //!
 //! The state contains a simple key-value store where each public key
 //! can store its associated encrypted credential so that the credential can be
 //! recovered by off-chain wallets. The `keys` for the store are the Ed25519
-//! public keys and the `values` are the `metadata + encrypted_credentials`. The
-//! metadata associated with the credential contains information needed for
-//! decoding the encrypted credential. Only if an off-chain tool has access to
-//! the AES secret key, it can decode the credentials in this smart contract
-//! (credentials cannot be decoded by third parties to preserve privacy). To
-//! enter credentials into this contract, only the associated private key to the
-//! public key can authorize the creation of an entry in this smart contract for
-//! its key. To authorize the entry the associated private key to the public key
-//! signs its metadata/encrypted_credential that it wants to store. This ensures
-//! that entries are generated/authorized by the public/private key pair
-//! that they are associated with. The entries in this contract are immutable
-//! and cannot be updated anymore once stored in this contract. Each
-//! public/private key pair can authorize/generate up to one entry in this
-//! contract.
+//! public keys and the `values` are the `version + encrypted_credentials`. The
+//! version information is used by off-chain tools for decoding the encrypted
+//! credential. Only if an off-chain tool has access to the AES secret key, it
+//! can decode the credentials in this smart contract (credentials cannot be
+//! decoded by third parties to preserve privacy). To enter credentials into
+//! this contract, only the associated private key to the public key can
+//! authorize the creation of an entry in this smart contract for its key. To
+//! authorize the entry the associated private key to the public key signs its
+//! version/ encrypted_credential that it wants to store. This ensures that
+//! entries are generated/authorized by the public/private key pair that they
+//! are associated with. The entries in this contract are immutable and cannot
+//! be updated anymore once stored in this contract. Each public/private key
+//! pair can authorize/generate up to one entry in this contract.
 //!
 //! The Concordium wallets generate a new public-private key pair and an
 //! AES-256 secret key from a key path (example/dummy key path m/1’/2’/3’/4/0)
@@ -51,16 +49,15 @@ use concordium_std::*;
 
 const SIGNATURE_DOMAIN: &str = "WEB3ID:STORE";
 
-/// Metadata is a string of two bytes that contain information needed for
-/// decoding the credentials.
-#[derive(Serial, SchemaType, Clone, Copy, Deserial, PartialEq, Debug)]
-struct Metadata([u8; 2]);
+/// Version is a number that is used when decoding a credential by off-chain
+/// tools.
+type Version = u16;
 
 /// Part of the contract state.
 #[derive(Serial, Deserial, Clone)]
 struct CredentialState {
-    /// Metadata associated with the credential.
-    metadata:             Metadata,
+    /// Version of the credential.
+    version:              Version,
     /// The `encrypted_credential` stored in this contract.
     encrypted_credential: Vec<u8>,
 }
@@ -122,12 +119,12 @@ impl From<LogError> for CustomContractError {
 }
 
 /// Credential state that is used as the input parameter or the return value for
-/// the `store/view` contract functions. It contains the metadata and the
+/// the `store/view` contract functions. It contains the version and the
 /// encrypted_credential.
 #[derive(Serialize, SchemaType, PartialEq)]
 struct SimpleCredentialState {
-    /// Metadata associated with the credential.
-    metadata:             Metadata,
+    /// Version of the credential..
+    version:              Version,
     /// The `encrypted_credential`.
     encrypted_credential: Vec<u8>,
 }
@@ -175,7 +172,7 @@ fn view<S: HasStateApi>(
 
     Ok(host.state().credential_registry.get(&public_key).map(|credential_state| {
         SimpleCredentialState {
-            metadata:             credential_state.metadata,
+            version:              credential_state.version,
             encrypted_credential: credential_state.encrypted_credential.clone(),
         }
     }))
@@ -184,51 +181,38 @@ fn view<S: HasStateApi>(
 /// The parameter type for the contract function `store`.
 #[derive(Serialize, SchemaType, Debug)]
 pub struct StoreParam {
-    /// The contract_address that the signature is intended for.
-    contract_address:     ContractAddress,
-    /// The serialized encrypted_credential.
-    #[concordium(size_length = 2)]
-    encrypted_credential: Vec<u8>,
-    /// Metadata associated with the credential.
-    metadata:             Metadata,
     /// Public key that created the above signature.
-    public_key:           PublicKeyEd25519,
+    public_key: PublicKeyEd25519,
     /// Signature.
-    signature:            SignatureEd25519,
-    /// A timestamp to make signatures expire.
-    timestamp:            Timestamp,
+    signature:  SignatureEd25519,
+    data:       DataToSign,
 }
 
 impl StoreParam {
     /// Prepare the message bytes for signature verification
     fn message_bytes(&self, bytes: &mut Vec<u8>) -> ContractResult<()> {
-        self.contract_address.serial(bytes).map_err(|_| CustomContractError::SerializationError)?;
-        self.encrypted_credential
-            .serial::<Vec<_>>(bytes)
-            .map_err(|_| CustomContractError::SerializationError)?;
-        self.metadata.serial(bytes).map_err(|_| CustomContractError::SerializationError)?;
-        self.timestamp.serial(bytes).map_err(|_| CustomContractError::SerializationError)?;
+        self.data.serial(bytes).map_err(|_| CustomContractError::SerializationError)?;
         Ok(())
     }
 }
 
 /// The parameter type for the contract function `serializationHelper`.
-#[derive(Serialize, SchemaType)]
-pub struct SerializationHelperParam {
+#[derive(Serialize, SchemaType, Debug)]
+pub struct DataToSign {
+    /// A timestamp to make signatures expire.
+    timestamp:            Timestamp,
     /// The contract_address that the signature is intended for.
     contract_address:     ContractAddress,
+    /// Version of the credential.
+    version:              Version,
     /// The serialized encrypted_credential.
     #[concordium(size_length = 2)]
     encrypted_credential: Vec<u8>,
-    /// Metadata associated with the credential.
-    metadata:             Metadata,
-    /// A timestamp to make signatures expire.
-    timestamp:            Timestamp,
 }
 
 /// Helper function that can be invoked at the front end to serialize
-/// the `SerializationHelperParam` to be signed by the wallet. The
-/// `SerializationHelperParam` includes all the input parameters from
+/// the `DataToSign` to be signed by the wallet. The
+/// `DataToSign` includes all the input parameters from
 /// `StoreParam` except for the `signature` and the `public_key`. We only need
 /// the input parameter schema of this function at the front end. The
 /// `serializationHelper` function is not executed at any point in time,
@@ -236,7 +220,7 @@ pub struct SerializationHelperParam {
 #[receive(
     contract = "credential-registry-storage",
     name = "serializationHelper",
-    parameter = "SerializationHelperParam"
+    parameter = "DataToSign"
 )]
 fn contract_serialization_helper<S: HasStateApi>(
     _ctx: &impl HasReceiveContext,
@@ -273,10 +257,10 @@ fn store<S: HasStateApi>(
     let param: StoreParam = ctx.parameter_cursor().get()?;
 
     // Check that the signature was intended for this contract.
-    ensure_eq!(param.contract_address, ctx.self_address(), CustomContractError::WrongContract);
+    ensure_eq!(param.data.contract_address, ctx.self_address(), CustomContractError::WrongContract);
 
     // Check signature is not expired.
-    ensure!(param.timestamp >= ctx.metadata().slot_time(), CustomContractError::Expired);
+    ensure!(param.data.timestamp >= ctx.metadata().slot_time(), CustomContractError::Expired);
 
     // Perepare message bytes as it is signed by the wallet.
     // Note that the message is prepended by a domain separation string.
@@ -290,8 +274,8 @@ fn store<S: HasStateApi>(
     );
 
     let entry = host.state_mut().credential_registry.insert(param.public_key, CredentialState {
-        metadata:             param.metadata,
-        encrypted_credential: param.encrypted_credential,
+        version:              param.data.version,
+        encrypted_credential: param.data.encrypted_credential,
     });
 
     ensure!(entry.is_none(), CustomContractError::CredentialAlreadyRegisteredForGivenPublicKey);
@@ -317,13 +301,13 @@ mod tests {
         115, 6, 164, 14, 89, 135, 129, 114, 208, 90, 66, 99,
     ]);
     const SIGNATURE: SignatureEd25519 = SignatureEd25519([
-        79, 193, 117, 35, 254, 245, 122, 77, 3, 247, 197, 145, 79, 59, 133, 196, 111, 160, 34, 85,
-        70, 193, 100, 197, 117, 128, 233, 90, 93, 97, 88, 62, 151, 197, 114, 149, 143, 213, 86,
-        166, 64, 251, 17, 34, 192, 177, 39, 53, 93, 221, 86, 51, 171, 133, 248, 120, 97, 176, 99,
-        133, 64, 173, 17, 10,
+        218, 87, 238, 248, 169, 170, 187, 125, 142, 60, 95, 218, 221, 193, 157, 184, 243, 243, 254,
+        38, 162, 26, 5, 126, 58, 69, 15, 236, 141, 16, 196, 8, 187, 94, 139, 108, 249, 180, 135,
+        74, 157, 80, 125, 18, 250, 220, 162, 26, 2, 87, 193, 248, 194, 216, 19, 6, 173, 4, 135,
+        227, 62, 69, 83, 7,
     ]);
     const ENCRYPTED_CREDENTIAL: [u8; 2] = [43, 1];
-    const METADATA: Metadata = Metadata([43, 1]);
+    const VERSION: Version = 0;
 
     // Test initialization succeeds.
     #[concordium_test]
@@ -362,17 +346,21 @@ mod tests {
         });
         ctx.set_metadata_slot_time(Timestamp::from_timestamp_millis(0));
 
-        // Set up the parameter.
-        let parameter = StoreParam {
-            signature:            SIGNATURE,
-            public_key:           PUBLIC_KEY,
+        let data = DataToSign {
             contract_address:     ContractAddress {
                 index:    0,
                 subindex: 0,
             },
             timestamp:            Timestamp::from_timestamp_millis(10000000000),
-            metadata:             METADATA,
+            version:              VERSION,
             encrypted_credential: ENCRYPTED_CREDENTIAL.to_vec(),
+        };
+
+        // Set up the parameter.
+        let parameter = StoreParam {
+            signature: SIGNATURE,
+            public_key: PUBLIC_KEY,
+            data,
         };
 
         let parameter_bytes = to_bytes(&parameter);
@@ -384,7 +372,7 @@ mod tests {
         let result = store(&ctx, &mut host, &mut logger, &crypto_primitives);
 
         // Check the result.
-        claim!(result.is_ok(), "Results in rejection");
+        claim!(result.is_ok(), "Results in rejection: {:?}", result);
 
         // Check the logs.
         claim_eq!(logger.logs.len(), 1, "One event should be logged");
@@ -407,7 +395,7 @@ mod tests {
         claim_eq!(
             result.expect_report("Expect credential as return value"),
             Some(SimpleCredentialState {
-                metadata:             METADATA,
+                version:              VERSION,
                 encrypted_credential: ENCRYPTED_CREDENTIAL.to_vec(),
             }),
             "Credential should be viewable"